requirements.txt
transformers
torch
pandas
transformers
torch
PyPDF2
python-docx